+++ eras/byron/ledger/impl/src/Cardano/Chain/Update/SystemTag.hs
@@ -97,7 +97,7 @@ instance DecCBOR SystemTagError where
     case tag of
       0 -> checkSize 2 >> SystemTagNotAscii <$> decCBOR
       1 -> checkSize 2 >> SystemTagTooLong <$> decCBOR
-      _ -> cborError $ DecoderErrorUnknownTag "SystemTagError" tag
+      _ -> cborError $ DecoderErrorUnknownTag "SystemTagError" (fromIntegral tag)
 
 instance B.Buildable SystemTagError where
   build = \case
